+++ src/lib/converter/jsdoc.ts
@@ -3,7 +3,7 @@
 // @callback
 
 import { ok } from "assert";
-import * as ts from "typescript";
+import ts from "typescript";
 import {
     DeclarationReflection,
     IntrinsicType,