+++ src/lib/converter/types.ts
@@ -1,5 +1,5 @@
 import assert from "assert";
-import * as ts from "typescript";
+import ts from "typescript";
 import {
     ArrayType,
     ConditionalType,