+++ src/test/converter2/behavior/excludeNotDocumentedKinds.ts
@@ -1,0 +1,11 @@
+export interface NotDoc {
+    /** Doc */
+    prop: 123;
+
+    notDoc: 456;
+}
+
+/** Doc */
+export function identity<T>(x: T): T {
+    return x;
+}