+++ src/test/converter2/issues/gh2135.ts
@@ -1,0 +1,7 @@
+export class Camera {
+    /** One */
+    static useCameraPermissions = createPermissionHook();
+}
+
+/** Two */
+declare function createPermissionHook(): () => void;