+++ src/test/issueTests.ts
@@ -29,8 +29,8 @@ function query(project: ProjectReflection, name: string) {
 }
 
 export const issueTests: {
-    [issue: `pre${string}`]: (app: Application) => void;
-    [issue: `gh${string}`]: (
+    [issue: `pre${bigint}${string}`]: (app: Application) => void;
+    [issue: `gh${bigint}${string}`]: (
         project: ProjectReflection,
         logger: TestLogger
     ) => void;
@@ -819,5 +819,34 @@ export const issueTests: {
             Comment.combineDisplayParts(x.comment?.summary),
             "Foo type comment"
         );
+    },
+
+    gh2135(project) {
+        const hook = query(project, "Camera.useCameraPermissions");
+        equal(hook.type?.type, "reflection" as const);
+        equal(
+            Comment.combineDisplayParts(
+                hook.type.declaration.signatures![0].comment?.summary
+            ),
+            "One"
+        );
+    },
+
+    pre2156(app) {
+        app.options.setValue("excludeNotDocumented", true);
+    },
+    gh2156(project) {
+        const foo = query(project, "foo");
+        equal(foo.signatures?.length, 1);
+        equal(
+            Comment.combineDisplayParts(foo.signatures[0].comment?.summary),
+            "Is documented"
+        );
+    },
+
+    gh2175(project) {
+        const def = query(project, "default");
+        equal(def.type?.type, "intrinsic");
+        equal(def.type.toString(), "undefined");
     },
 };